Move artifact reading to a separate file.
Changelog: None
