+++ cli/mender-artifact/artifacts.go
@@ -100,89 +100,6 @@ func read(ar *areader.Reader, verify areader.SignatureVerifyFn,
 	return ar, nil
 }
 
-func readArtifact(c *cli.Context) error {
-	if c.NArg() == 0 {
-		return cli.NewExitError("Nothing specified, nothing read. \nMaybe you wanted"+
-			" to say 'artifacts read <pathspec>'?", 1)
-	}
-
-	f, err := os.Open(c.Args().First())
-	if err != nil {
-		return cli.NewExitError("Can not open '"+c.Args().First()+"' file.", 1)
-	}
-	defer f.Close()
-
-	var verifyCallback areader.SignatureVerifyFn
-
-	if len(c.String("key")) != 0 {
-		key, keyErr := getKey(c.String("key"))
-		if keyErr != nil {
-			return cli.NewExitError(keyErr.Error(), 1)
-		}
-		s := artifact.NewVerifier(key)
-		verifyCallback = s.Verify
-	}
-
-	// if key is not provided just continue reading artifact returning
-	// info that signature can not be verified
-	sigInfo := "no signature"
-	ver := func(message, sig []byte) error {
-		sigInfo = "signed but no key for verification provided; " +
-			"please use `-k` option for providing verification key"
-		if verifyCallback != nil {
-			err = verifyCallback(message, sig)
-			if err != nil {
-				sigInfo = "signed; verification using provided key failed"
-			} else {
-				sigInfo = "signed and verified correctly"
-			}
-		}
-		return nil
-	}
-
-	var scripts []string
-	readScripts := func(r io.Reader, info os.FileInfo) error {
-		scripts = append(scripts, info.Name())
-		return nil
-	}
-
-	ar := areader.NewReader(f)
-	r, err := read(ar, ver, readScripts)
-	if err != nil {
-		return cli.NewExitError(err.Error(), 1)
-	}
-
-	inst := r.GetHandlers()
-	info := r.GetInfo()
-
-	fmt.Printf("Mender artifact:\n")
-	fmt.Printf("  Name: %s\n", r.GetArtifactName())
-	fmt.Printf("  Format: %s\n", info.Format)
-	fmt.Printf("  Version: %d\n", info.Version)
-	fmt.Printf("  Signature: %s\n", sigInfo)
-	fmt.Printf("  Compatible devices: '%s'\n", r.GetCompatibleDevices())
-	if len(scripts) > 0 {
-		fmt.Printf("  State scripts:\n")
-	}
-	for _, scr := range scripts {
-		fmt.Printf("    %s\n", scr)
-	}
-	fmt.Printf("\nUpdates:\n")
-
-	for k, p := range inst {
-		fmt.Printf("  %04d:\n", k)
-		fmt.Printf("    Type:   %s\n", p.GetType())
-		for _, f := range p.GetUpdateFiles() {
-			fmt.Printf("    Files:\n")
-			fmt.Printf("      name:     %s\n", f.Name)
-			fmt.Printf("      size:     %d\n", f.Size)
-			fmt.Printf("      modified: %s\n", f.Date)
-			fmt.Printf("      checksum: %s\n", f.Checksum)
-		}
-	}
-	return nil
-}
-
 func getKey(path string) ([]byte, error) {
 	f, err := os.Open(path)
 	if err != nil {
